[k8s] Implement an operator for a KuduCluster CRD
The operator is a custom controller that monitors and maintains the
correct Kudu Kubernetes cluster state (a StatefulSet of Kudu masters,
a StatefulSet of Kudu tservers, and Services for the masters, tservers,
the master UI). It can be launched as Go program outside the cluster
(best for local development) or as a Kubernetes Deployment inside the
cluster. Currently, the operator maintains the specified number of
masters and tservers, and triggers the Kudu rebalancer if tservers are
added or removed.

My notes on Kubernetes and Operators here:
https://docs.google.com/document/d/1zre8LLq6C0Dh64o16rNknBfp6wwtpytRySXSzUt6prk

Change-Id: I1c855f3bab7bbd501f6237a23781f7ae200ba3db